requirements.txt
python-dotenv == 0.0.5
requests == 2.32.3